Add files via upload
After data-cleaning and pre processing. I used Vader which is a Sentiment Analysis tool giving scores to the words within a review, which allowed me to determine a positive or negative score for each review. I further built a Matrix with Sentiment and Temporal Analysis, to identify patterns and correlations between suspicious users and reviews within my dataset.